Time: 0 ms (100.00%), Space: 5.9 MB (67.26%) - LeetHub

+++ 374-guess-number-higher-or-lower/374-guess-number-higher-or-lower.cpp
@@ -10,11 +10,11 @@
 class Solution {
 public:
     int guessNumber(int n) {
-        long long low=1;
-        long long high=n;
+        long  low=1;
+        long  high=n;
         while(low<=high)
         {
-            long long mid=(low+high)/2;
+            long mid=(low+high)/2;
             if(guess(mid)==0) return mid;
             else if(guess(mid)==1) low=mid+1;
             else high=mid-1;
